coursework_5_2/pyproject.toml
[tool.poetry]
name = "backend-django"
version = "0.1.0"
description = ""
authors = ["Julia-Gareeva <uliastekolnikova12822@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djoser = "^2.1.0"
django = "^4.1.4"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"
drf-spectacular = "^0.24.2"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
